Purge all the things and clean up
