--- module/input/SensorFilter/data/config/FootDownNetwork.yaml
@@ -1,169 +1,0 @@
-filter:
-  certainty_threshold: 0.5
-  noise_factor: 1e-4
-  uncertainty_threshold: 0.5
-network:
-  input:
-    accelerometer: true
-    fields:
-      - LOAD
-      - VELOCITY
-    gyroscope: false
-    servos:
-      - HIP_PITCH
-      - KNEE
-      - ANKLE_PITCH
-  layers:
-    - biases:
-        - 0.4433150589466095
-        - -0.21458734571933746
-        - -0.665310800075531
-        - 0.7759745717048645
-        - 0.9789865612983704
-        - 0.4197600781917572
-        - -0.9810238480567932
-        - 0.19538094103336334
-      weights:
-        - - -1.1843377351760864
-          - -0.6812138557434082
-          - -3.444648027420044
-          - -5.709575653076172
-          - -2.1864359378814697
-          - -2.8936030864715576
-          - 4.4670538902282715
-          - -3.7462191581726074
-        - - 1.000468134880066
-          - -2.3140757083892822
-          - 0.9452422857284546
-          - 1.7822060585021973
-          - 0.09961029887199402
-          - 0.21727101504802704
-          - -1.3409172296524048
-          - 1.29275643825531
-        - - 0.8036074638366699
-          - 7.460040092468262
-          - -1.7128444910049438
-          - 1.691080093383789
-          - 0.7580970525741577
-          - 7.362337112426758
-          - -4.676390171051025
-          - 3.283323049545288
-        - - -0.11287917196750641
-          - -1.8725136518478394
-          - -0.07176670432090759
-          - -0.4219309985637665
-          - -0.057131197303533554
-          - -2.063908815383911
-          - 1.582884430885315
-          - -0.7892264723777771
-        - - 5.786834239959717
-          - 3.3275692462921143
-          - -0.8779624700546265
-          - 4.053295612335205
-          - -2.160700798034668
-          - 2.399670362472534
-          - 3.247880697250366
-          - 0.39634084701538086
-        - - -0.5917699337005615
-          - -0.6622474193572998
-          - 0.13019977509975433
-          - -1.143958568572998
-          - 0.22768908739089966
-          - 0.06266005337238312
-          - -0.16209638118743896
-          - -0.14163614809513092
-        - - 0.9591425061225891
-          - -3.6435444355010986
-          - 0.6769518256187439
-          - 1.0733160972595215
-          - -4.838194370269775
-          - 1.7473992109298706
-          - -5.438755989074707
-          - 2.0149714946746826
-        - - -0.49153536558151245
-          - 1.0969140529632568
-          - -0.574046790599823
-          - -0.14157184958457947
-          - 1.3451273441314697
-          - -0.5798219442367554
-          - 0.861062228679657
-          - -0.08049441128969193
-        - - -2.8889219760894775
-          - 2.1305079460144043
-          - 2.16501522064209
-          - -7.052975177764893
-          - -0.8476078510284424
-          - 3.261120080947876
-          - -0.04408276453614235
-          - -1.9315909147262573
-        - - 4.288317680358887
-          - -1.2079801559448242
-          - -0.4645354151725769
-          - 0.8387446403503418
-          - 0.305756151676178
-          - 2.380720615386963
-          - -0.3690907061100006
-          - -1.2014425992965698
-        - - 1.4049508571624756
-          - 8.983287811279297
-          - 3.6116161346435547
-          - 1.6626759767532349
-          - 3.482375383377075
-          - 4.681126594543457
-          - -0.40511631965637207
-          - -0.834507942199707
-        - - -2.123495101928711
-          - 1.6786580085754395
-          - -1.3272532224655151
-          - 0.1453731209039688
-          - 0.07116696983575821
-          - -1.7513643503189087
-          - 0.6509838700294495
-          - 2.0158207416534424
-        - - -0.16732151806354523
-          - -0.2367517650127411
-          - 0.9195804595947266
-          - 0.06437806040048599
-          - -0.19833320379257202
-          - -0.08134431391954422
-          - -0.15488995611667633
-          - -0.11292699724435806
-        - - 0.07729703933000565
-          - 0.07630302011966705
-          - -0.067111536860466
-          - -0.0159927885979414
-          - 0.07020669430494308
-          - 0.08038739860057831
-          - -0.22290757298469543
-          - -0.015449780970811844
-        - - 0.12052829563617706
-          - -0.016871333122253418
-          - 0.08979465812444687
-          - 0.04066053405404091
-          - -0.1047956570982933
-          - 0.047791026532649994
-          - 0.11397455632686615
-          - 0.06255297362804413
-    - biases:
-        - 0.9622185230255127
-        - 6.901426792144775
-      weights:
-        - - -5.303719997406006
-          - 0.02708369679749012
-        - - -1.0492877960205078
-          - 3.5611987113952637
-        - - -3.416400909423828
-          - -3.6899161338806152
-        - - 3.9959511756896973
-          - 1.2409859895706177
-        - - -2.8959169387817383
-          - -5.379881858825684
-        - - 4.792142391204834
-          - -1.6149020195007324
-        - - 0.23115263879299164
-          - -8.717751502990723
-        - - -1.2211687564849854
-          - -5.894148349761963
-train:
-  foot_delta: 0.005
-  lr_duplicate: true
