+++ module/input/SensorFilter/src/SensorFilter.hpp
@@ -34,60 +34,57 @@
 #include <tinyrobotics/kinematics.hpp>
 #include <tinyrobotics/parser.hpp>
 
-#include "MotionModel.hpp"
-#include "VirtualLoadSensor.hpp"
-
 #include "extension/Configuration.hpp"
 
-#include "message/behaviour/state/Stability.hpp"
+#include "message/actuation/BodySide.hpp"
 #include "message/behaviour/state/WalkState.hpp"
 #include "message/input/Sensors.hpp"
+#include "message/localisation/Field.hpp"
 #include "message/platform/RawSensors.hpp"
 
-#include "utility/math/filter/KalmanFilter.hpp"
-#include "utility/math/filter/UKF.hpp"
-
-using extension::Configuration;
+#include "utility/actuation/tinyrobotics.hpp"
+#include "utility/input/FrameID.hpp"
+#include "utility/input/LimbID.hpp"
+#include "utility/input/ServoID.hpp"
+#include "utility/math/euler.hpp"
+#include "utility/math/filter/MahonyFilter.hpp"
+#include "utility/nusight/NUhelpers.hpp"
+#include "utility/platform/RawSensors.hpp"
+#include "utility/support/yaml_expression.hpp"
 
 namespace module::input {
 
-    using message::behaviour::state::Stability;
+    using utility::actuation::tinyrobotics::forward_kinematics_to_servo_map;
+    using utility::actuation::tinyrobotics::sensors_to_configuration;
+    using utility::input::FrameID;
+    using utility::input::ServoID;
+    using utility::math::euler::EulerIntrinsicToMatrix;
+    using utility::math::euler::MatrixToEulerIntrinsic;
+    using utility::nusight::graph;
+    using utility::platform::getRawServo;
+    using utility::platform::make_packet_error_string;
+    using utility::platform::make_servo_hardware_error_string;
+    using utility::support::Expression;
+
+    using message::actuation::BodySide;
     using message::behaviour::state::WalkState;
     using message::input::Sensors;
+    using message::localisation::ResetFieldLocalisation;
+    using message::platform::ButtonLeftDown;
+    using message::platform::ButtonLeftUp;
+    using message::platform::ButtonMiddleDown;
+    using message::platform::ButtonMiddleUp;
     using message::platform::RawSensors;
 
-    /**
-     * @author Jade Fountain
-     * @author Trent Houliston
-     */
+    using tinyrobotics::forward_kinematics;
+
     class SensorFilter : public NUClear::Reactor {
     public:
         explicit SensorFilter(std::unique_ptr<NUClear::Environment> environment);
 
     private:
-        /// @brief Number of actuatable joints in the NUgus robot
-        static const int n_joints = 20;
-
-        /// @brief tinyrobotics NUgus model used for kinematics
-        tinyrobotics::Model<double, n_joints> nugus_model;
-
-        /// @brief Unscented kalman filter for pose estimation
-        utility::math::filter::UKF<double, MotionModel> ukf{};
-
-        /// @brief Number of states in the Kalman filter. x = [roll, pitch, roll rate, pitch rate]'
-        static const size_t n_states = 4;
-
-        /// @brief Number of inputs in the Kalman filter. Model has no inputs
-        static const size_t n_inputs = 0;
-
-        /// @brief Number of measurements in the Kalman filter. y = [roll, pitch, roll rate, pitch rate]'
-        static const size_t n_measurements = 4;
-
-        /// @brief Kalman filter for pose estimation
-        utility::math::filter::KalmanFilter<double, n_states, n_inputs, n_measurements> kf{};
-
         struct FootDownMethod {
-            enum Value { UNKNOWN = 0, Z_HEIGHT = 1, LOAD = 2, FSR = 3 };
+            enum Value { UNKNOWN = 0, Z_HEIGHT = 1, FSR = 2 };
             Value value = Value::UNKNOWN;
 
             // Constructors
@@ -97,7 +94,6 @@ namespace module::input {
             FootDownMethod(std::string const& str) {
                 // clang-format off
                         if      (str == "Z_HEIGHT") { value = Value::Z_HEIGHT; }
-                        else if (str == "LOAD") { value = Value::LOAD; }
                         else if (str == "FSR")  { value = Value::FSR; }
                         else {
                             value = Value::UNKNOWN;
@@ -113,63 +109,13 @@ namespace module::input {
             [[nodiscard]] operator std::string() const {
                 switch (value) {
                     case Value::Z_HEIGHT: return "Z_HEIGHT";
-                    case Value::LOAD: return "VIRTUAL";
                     case Value::FSR: return "FSR";
                     default: throw std::runtime_error("enum Method's value is corrupt, unknown value stored");
                 }
             }
         };
 
-        struct FilteringMethod {
-            enum Value { UNKNOWN = 0, UKF = 1, KF = 2, MAHONY = 3, GROUND_TRUTH = 4 };
-            Value value = Value::UNKNOWN;
-
-            // Constructors
-            FilteringMethod() = default;
-            FilteringMethod(int const& v) : value(static_cast<Value>(v)) {}
-            FilteringMethod(Value const& v) : value(v) {}
-            FilteringMethod(std::string const& str) {
-                // clang-format off
-                        if      (str == "UKF") { value = Value::UKF; }
-                        else if (str == "KF") { value = Value::KF; }
-                        else if (str == "MAHONY")  { value = Value::MAHONY; }
-                        else if (str == "GROUND_TRUTH")  { value = Value::GROUND_TRUTH; }
-                        else {
-                            value = Value::UNKNOWN;
-                            throw std::runtime_error("String " + str + " did not match any enum for FilteringMethod");
-                        }
-                // clang-format on
-            }
-
-            // Conversions
-            [[nodiscard]] operator Value() const {
-                return value;
-            }
-            [[nodiscard]] operator std::string() const {
-                switch (value) {
-                    case Value::UKF: return "UKF";
-                    case Value::KF: return "KF";
-                    case Value::MAHONY: return "MAHONY";
-                    case Value::GROUND_TRUTH: return "GROUND_TRUTH";
-                    default: throw std::runtime_error("enum Method's value is corrupt, unknown value stored");
-                }
-            }
-        };
-
-
         struct Config {
-            /// @brief Path to NUgus URDF file
-            std::string urdf_path = "";
-
-            /// @brief Initial transform from torso {t} to world {w} space
-            Eigen::Isometry3d initial_Hwt = Eigen::Isometry3d::Identity();
-
-            /// @brief Config for the button debouncer
-            struct Button {
-                Button() = default;
-                /// @brief The number of times a button must be pressed before it is considered pressed
-                int debounce_threshold = 0;
-            } buttons{};
 
             /// @brief Config for the foot down detector
             struct FootDown {
@@ -193,84 +139,16 @@ namespace module::input {
                 FootDownMethod current_method                        = FootDownMethod::Z_HEIGHT;
                 std::map<FootDownMethod, float> certainty_thresholds = {
                     {FootDownMethod::Z_HEIGHT, 0.01f},
-                    {FootDownMethod::LOAD, 0.05f},
                     {FootDownMethod::FSR, 60.0f},
                 };
-            } footDown;
+            } foot_down;
 
-            /// @brief Specifies the filtering method to use (UKF, KF, MAHONY)
-            FilteringMethod filtering_method;
+            /// @brief The number of times a button must be pressed before it is considered pressed
+            int button_debounce_threshold = 0;
 
-            //  **************************************** UKF Config ****************************************
-            /// @brief Config for the UKF
-            struct UKF {
-                Eigen::Vector3d velocity_decay = Eigen::Vector3d::Zero();
+            /// @brief Initial transform from torso {t} to world {w} space
+            Eigen::Isometry3d initial_Hwt = Eigen::Isometry3d::Identity();
 
-                struct Noise {
-                    Noise() = default;
-                    struct Measurement {
-                        Eigen::Matrix3d accelerometer           = Eigen::Matrix3d::Zero();
-                        Eigen::Matrix3d accelerometer_magnitude = Eigen::Matrix3d::Zero();
-                        Eigen::Matrix3d gyroscope               = Eigen::Matrix3d::Zero();
-                        Eigen::Matrix3d flat_foot_odometry      = Eigen::Matrix3d::Zero();
-                        Eigen::Matrix4d flat_foot_orientation   = Eigen::Matrix4d::Zero();
-                    } measurement{};
-                    struct Process {
-                        Eigen::Vector3d position            = Eigen::Vector3d::Zero();
-                        Eigen::Vector3d velocity            = Eigen::Vector3d::Zero();
-                        Eigen::Vector4d rotation            = Eigen::Vector4d::Zero();
-                        Eigen::Vector3d rotational_velocity = Eigen::Vector3d::Zero();
-                        Eigen::Vector3d gyroscope_bias      = Eigen::Vector3d::Zero();
-                    } process{};
-                } noise{};
-                struct Initial {
-                    Initial() = default;
-                    struct Mean {
-                        Eigen::Vector3d position            = Eigen::Vector3d::Zero();
-                        Eigen::Vector3d velocity            = Eigen::Vector3d::Zero();
-                        Eigen::Vector4d rotation            = Eigen::Vector4d::Zero();
-                        Eigen::Vector3d rotational_velocity = Eigen::Vector3d::Zero();
-                        Eigen::Vector3d gyroscope_bias      = Eigen::Vector3d::Zero();
-                    } mean{};
-                    struct Covariance {
-                        Eigen::Vector3d position            = Eigen::Vector3d::Zero();
-                        Eigen::Vector3d velocity            = Eigen::Vector3d::Zero();
-                        Eigen::Vector4d rotation            = Eigen::Vector4d::Zero();
-                        Eigen::Vector3d rotational_velocity = Eigen::Vector3d::Zero();
-                        Eigen::Vector3d gyroscope_bias      = Eigen::Vector3d::Zero();
-                    } covariance{};
-                } initial{};
-            } ukf{};
-
-            /// @brief Initial state of the for the UKF filter
-            MotionModel<double>::StateVec initial_mean;
-
-            /// @brief Initial covariance of the for the UKF filter
-            MotionModel<double>::StateVec initial_covariance;
-
-            /// @brief Parameter for scaling the walk command to better match actual achieved velocity
-            Eigen::Vector3d deadreckoning_scale = Eigen::Vector3d::Zero();
-
-            /// @brief Initial frame for anchoring deadreckoning
-            std::string initial_anchor_frame = "";
-
-            //  **************************************** Kalman Filter Config ****************************************
-            /// @brief Kalman Continuous time process model
-            Eigen::Matrix<double, n_states, n_states> Ac;
-
-            /// @brief Kalman Continuous time input model
-            Eigen::Matrix<double, n_states, n_inputs> Bc;
-
-            /// @brief Kalman Continuous time measurement model
-            Eigen::Matrix<double, n_measurements, n_states> C;
-
-            /// @brief Kalman Process noise
-            Eigen::Matrix<double, n_states, n_states> Q;
-
-            /// @brief Kalman Measurement noise
-            Eigen::Matrix<double, n_measurements, n_measurements> R;
-
-            //  **************************************** Mahony Filter Config ****************************************
             /// @brief Mahony filter bias
             Eigen::Vector3d initial_bias = Eigen::Vector3d::Zero();
 
@@ -280,6 +158,39 @@ namespace module::input {
             /// @brief Mahony filter integral gain
             double Kp = 0.0;
         } cfg;
+
+        /// @brief Number of actuatable joints in the NUgus robot
+        static const int n_servos = 20;
+
+        /// @brief tinyrobotics model of NUgus used for kinematics
+        tinyrobotics::Model<double, n_servos> nugus_model;
+
+        /// @brief Current support phase of the robot
+        WalkState::SupportPhase current_support_phase = WalkState::SupportPhase::LEFT;
+
+        /// @brief Transform from anchor {a} to world {w} space
+        Eigen::Isometry3d Hwa = Eigen::Isometry3d::Identity();
+
+        /// @brief Transform from torso {t} to anchor {a} space
+        Eigen::Isometry3d Hat = Eigen::Isometry3d::Identity();
+
+        /// @brief Transform from torso {t} to world {w} space using the mahony filter
+        Eigen::Isometry3d Hwt_mahony = Eigen::Isometry3d::Identity();
+
+        /// @brief Transform from torso {t} to world {w} space using mahony + anchor method
+        Eigen::Isometry3d Hwt_anchor = Eigen::Isometry3d::Identity();
+
+        /// @brief Bias used in the mahony filter, updates with each mahony update
+        Eigen::Vector3d bias_mahony = Eigen::Vector3d::Zero();
+
+        /// @brief Current walk command
+        Eigen::Vector3d walk_command = Eigen::Vector3d::Zero();
+
+        /// @brief Current state of the left button
+        bool left_down = false;
+
+        /// @brief Current state of the middle button
+        bool middle_down = false;
 
         /// @brief Updates the sensors message with raw sensor data, including the timestamp, battery
         /// voltage, servo sensors, accelerometer, gyroscope, buttons, and LED.
@@ -299,116 +210,20 @@ namespace module::input {
         /// @param raw_sensors The raw sensor data
         void update_kinematics(std::unique_ptr<Sensors>& sensors, const RawSensors& raw_sensors);
 
-        /// @brief Runs a deadreckoning update on the odometry for x, y and yaw using the walk command
-        /// @param dt The time since the last update
-        /// @param walk_state Current state of walk engine
-        void integrate_walkcommand(const double dt, const Stability& stability, const WalkState& walk_state);
-
-        /// @brief Updates translational and yaw components of odometry using the anchor method
-        /// @param walk_state Current state of walk engine
-        void anchor_update(std::unique_ptr<Sensors>& sensors, const WalkState& walk_state);
-
-        /// @brief Configure UKF filter
-        void configure_ukf(const Configuration& config);
-
-        /// @brief Configure Kalman filter
-        void configure_kf(const Configuration& config);
-
-        /// @brief Configure Mahony filter
-        void configure_mahony(const Configuration& config);
-
-        /// @brief Reset the UKF state
-        void reset_ukf();
-
-        /// @brief Reset the Kalman filter state
-        void reset_kf();
-
-        /// @brief Reset the Mahony filter state
-        void reset_mahony();
-
-        /// @brief Updates the sensors message with odometry data filtered using UKF. This includes the
-        // position, orientation, velocity and rotational velocity of the torso in world space.
+        /// @brief Display debug information
         /// @param sensors The sensors message to update
-        /// @param previous_sensors The previous sensors message
         /// @param raw_sensors The raw sensor data
-        void update_odometry_ukf(std::unique_ptr<Sensors>& sensors,
-                                 const std::shared_ptr<const Sensors>& previous_sensors,
-                                 const RawSensors& raw_sensors);
-
-        /// @brief Updates the sensors message with odometry data filtered using Kalman Filter. This includes the
-        // position, orientation, velocity and rotational velocity of the torso in world space.
-        /// @param sensors The sensors message to update
-        /// @param previous_sensors The previous sensors message
-        /// @param raw_sensors The raw sensor data
-        void update_odometry_kf(std::unique_ptr<Sensors>& sensors,
-                                const std::shared_ptr<const Sensors>& previous_sensors,
-                                const RawSensors& raw_sensors,
-                                const std::shared_ptr<const Stability>& stability,
-                                const std::shared_ptr<const WalkState>& walk_state);
-
+        void debug_sensor_filter(std::unique_ptr<Sensors>& sensors, const RawSensors& raw_sensors);
 
         /// @brief Updates the sensors message with odometry data filtered using MahonyFilter. This includes the
         // position, orientation, velocity and rotational velocity of the torso in world space.
         /// @param sensors The sensors message to update
         /// @param previous_sensors The previous sensors message
         /// @param raw_sensors The raw sensor data
-        void update_odometry_mahony(std::unique_ptr<Sensors>& sensors,
-                                    const std::shared_ptr<const Sensors>& previous_sensors,
-                                    const RawSensors& raw_sensors,
-                                    const std::shared_ptr<const WalkState>& walk_state);
-
-        /// @brief Updates the sensors message with odometry data filtered using ground truth from WeBots. This includes
-        /// the position, orientation, velocity and rotational velocity of the torso in world space.
-        /// @param sensors The sensors message to update
-        /// @param previous_sensors The previous sensors message
-        /// @param raw_sensors The raw sensor data
-        void update_odometry_ground_truth(std::unique_ptr<Sensors>& sensors, const RawSensors& raw_sensors);
-
-        /// @brief Display debug information
-        /// @param sensors The sensors message to update
-        /// @param raw_sensors The raw sensor data
-        void debug_sensor_filter(std::unique_ptr<Sensors>& sensors, const RawSensors& raw_sensors);
-
-        /// @brief Transform from anchor {a} to world {w} space
-        Eigen::Isometry3d Hwa = Eigen::Isometry3d::Identity();
-
-        /// @brief Current support phase of the robot
-        WalkState::SupportPhase current_support_phase = WalkState::SupportPhase::LEFT;
-
-        /// @brief Dead reckoning yaw orientation of the robot in world space
-        double yaw = 0;
-
-        /// @brief Transform from torso {t} to world {w} space
-        Eigen::Isometry3d Hwt = Eigen::Isometry3d::Identity();
-
-        // @brief Transform from torso {t} to world {w} space using mahony filter (only roll and pitch estimation)
-        Eigen::Isometry3d Hwt_mahony = Eigen::Isometry3d::Identity();
-
-        /// @brief Bias used in the mahony filter, updates with each mahony update
-        Eigen::Vector3d bias_mahony = Eigen::Vector3d::Zero();
-
-        /// @brief Current walk command
-        Eigen::Vector3d walk_command = Eigen::Vector3d::Zero();
-
-        /// @brief Current state of the left button
-        bool left_down = false;
-
-        /// @brief Current state of the middle button
-        bool middle_down = false;
-
-        /// @brief Our sensor for foot down
-        VirtualLoadSensor<float> load_sensor{};
-
-        // This keeps track of whether each sides foot was down in the previous time step
-        // e.g. if right foot down at time t, then at time t+1, previous_foot_down[RightSide] = true
-        std::array<bool, 2> previous_foot_down = {false, false};
-
-        // Foot to world in foot-flat (both feet down) rotation at the timestep with the most recent foot landing
-        std::array<Eigen::Isometry3d, 2> footlanding_Hwf{};
-
-        // Handle for the sensor filter update loop, allows disabling new sensor updates when a reset event occurs
-        ReactionHandle update_loop{};
-        std::atomic_bool reset_filter{false};
+        void update_odometry(std::unique_ptr<Sensors>& sensors,
+                             const std::shared_ptr<const Sensors>& previous_sensors,
+                             const RawSensors& raw_sensors,
+                             const std::shared_ptr<const WalkState>& walk_state);
     };
 }  // namespace module::input
 #endif  // MODULES_INPUT_SENSORFILTER_HPP
